suit/transport: move common code to new file
