+++ sys/suit/transport/coap.c
@@ -22,162 +22,15 @@
  * @}
  */
 
-#include <assert.h>
-#include <inttypes.h>
-#include <string.h>
-
-#include "msg.h"
 #include "log.h"
+#include "suit.h"
 #include "net/nanocoap.h"
-#include "net/nanocoap_sock.h"
-#include "thread.h"
-#include "periph/pm.h"
-#include "ztimer.h"
-
-#ifdef MODULE_SUIT_TRANSPORT_COAP
 #include "suit/transport/coap.h"
-#include "net/sock/util.h"
-#endif
-#ifdef MODULE_SUIT_TRANSPORT_VFS
-#include "vfs_util.h"
-#endif
+#include "kernel_defines.h"
 
 #ifdef MODULE_RIOTBOOT_SLOT
 #include "riotboot/slot.h"
 #endif
-
-#ifdef MODULE_SUIT
-#include "suit.h"
-#include "suit/handlers.h"
-#include "suit/storage.h"
-#endif
-
-#if defined(MODULE_PROGRESS_BAR)
-#include "progress_bar.h"
-#endif
-
-#define ENABLE_DEBUG 0
-#include "debug.h"
-
-#ifndef SUIT_COAP_STACKSIZE
-/* allocate stack needed to do manifest validation */
-#define SUIT_COAP_STACKSIZE (3 * THREAD_STACKSIZE_LARGE)
-#endif
-
-#ifndef SUIT_COAP_PRIO
-#define SUIT_COAP_PRIO THREAD_PRIORITY_MAIN - 1
-#endif
-
-#ifndef SUIT_URL_MAX
-#define SUIT_URL_MAX            128
-#endif
-
-#ifndef SUIT_MANIFEST_BUFSIZE
-#define SUIT_MANIFEST_BUFSIZE   640
-#endif
-
-#define SUIT_MSG_TRIGGER        0x12345
-
-static char _stack[SUIT_COAP_STACKSIZE];
-static char _url[SUIT_URL_MAX];
-static uint8_t _manifest_buf[SUIT_MANIFEST_BUFSIZE];
-
-static kernel_pid_t _suit_coap_pid;
-
-static void _suit_handle_url(const char *url, coap_blksize_t blksize)
-{
-    ssize_t size;
-    LOG_INFO("suit_coap: downloading \"%s\"\n", url);
-
-    if (0) {}
-#ifdef MODULE_SUIT_TRANSPORT_COAP
-    else if (strncmp(url, "coap://", 7) == 0) {
-        size = nanocoap_get_blockwise_url_to_buf(url, blksize,
-                                                 _manifest_buf,
-                                                 sizeof(_manifest_buf));
-    }
-#endif
-#ifdef MODULE_SUIT_TRANSPORT_VFS
-    else if (strncmp(url, "file:/", 6) == 0) {
-        size = vfs_file_to_buffer(&url[6], _manifest_buf, sizeof(_manifest_buf));
-    }
-#endif
-    else {
-        LOG_WARNING("suit_coap: unsupported URL scheme!\n)");
-        return;
-    }
-
-    if (size >= 0) {
-        LOG_INFO("suit_coap: got manifest with size %u\n", (unsigned)size);
-
-#ifdef MODULE_SUIT
-        suit_manifest_t manifest;
-        memset(&manifest, 0, sizeof(manifest));
-
-        manifest.urlbuf = _url;
-        manifest.urlbuf_len = SUIT_URL_MAX;
-
-        int res;
-        if ((res = suit_parse(&manifest, _manifest_buf, size)) != SUIT_OK) {
-            LOG_INFO("suit_parse() failed. res=%i\n", res);
-            return;
-        }
-
-#endif
-#ifdef MODULE_SUIT_STORAGE_FLASHWRITE
-        if (res == 0) {
-            const riotboot_hdr_t *hdr = riotboot_slot_get_hdr(
-                riotboot_slot_other());
-            riotboot_hdr_print(hdr);
-            ztimer_sleep(ZTIMER_MSEC, 1 * MS_PER_SEC);
-
-            if (riotboot_hdr_validate(hdr) == 0) {
-                LOG_INFO("suit_coap: rebooting...\n");
-                pm_reboot();
-            }
-            else {
-                LOG_INFO("suit_coap: update failed, hdr invalid\n ");
-            }
-        }
-#endif
-    }
-    else {
-        LOG_INFO("suit_coap: error getting manifest\n");
-    }
-}
-
-static void *_suit_coap_thread(void *arg)
-{
-    (void)arg;
-
-    LOG_INFO("suit_coap: started.\n");
-    msg_t msg_queue[4];
-    msg_init_queue(msg_queue, 4);
-
-    _suit_coap_pid = thread_getpid();
-
-    msg_t m;
-    while (true) {
-        msg_receive(&m);
-        DEBUG("suit_coap: got msg with type %" PRIu32 "\n", m.content.value);
-        switch (m.content.value) {
-            case SUIT_MSG_TRIGGER:
-                LOG_INFO("suit_coap: trigger received\n");
-                _suit_handle_url(_url, CONFIG_SUIT_COAP_BLOCKSIZE);
-                break;
-            default:
-                LOG_WARNING("suit_coap: warning: unhandled msg\n");
-        }
-    }
-    return NULL;
-}
-
-void suit_coap_run(void)
-{
-    thread_create(_stack, SUIT_COAP_STACKSIZE, SUIT_COAP_PRIO,
-                  THREAD_CREATE_STACKTEST,
-                  _suit_coap_thread, NULL, "suit_coap");
-}
 
 static ssize_t _version_handler(coap_pkt_t *pkt, uint8_t *buf, size_t len,
                                 void *context)
@@ -186,25 +39,6 @@ static ssize_t _version_handler(coap_pkt_t *pkt, uint8_t *buf, size_t len,
     return coap_reply_simple(pkt, COAP_CODE_205, buf, len,
                              COAP_FORMAT_TEXT, (uint8_t *)"NONE", 4);
 }
-
-#ifdef MODULE_RIOTBOOT_SLOT
-static ssize_t _slot_handler(coap_pkt_t *pkt, uint8_t *buf, size_t len,
-                             void *context)
-{
-    /* context is passed either as NULL or 0x1 for /active or /inactive */
-    char c = '0';
-
-    if (context) {
-        c += riotboot_slot_other();
-    }
-    else {
-        c += riotboot_slot_current();
-    }
-
-    return coap_reply_simple(pkt, COAP_CODE_205, buf, len,
-                             COAP_FORMAT_TEXT, (uint8_t *)&c, 1);
-}
-#endif
 
 static ssize_t _trigger_handler(coap_pkt_t *pkt, uint8_t *buf, size_t len,
                                 void *context)
@@ -230,21 +64,31 @@ static ssize_t _trigger_handler(coap_pkt_t *pkt, uint8_t *buf, size_t len,
                              COAP_FORMAT_NONE, NULL, 0);
 }
 
-void suit_coap_trigger(const uint8_t *url, size_t len)
+#ifdef MODULE_RIOTBOOT_SLOT
+static ssize_t _slot_handler(coap_pkt_t *pkt, uint8_t *buf, size_t len,
+                             void *context)
 {
-    memcpy(_url, url, len);
-    _url[len] = '\0';
-    msg_t m = { .content.value = SUIT_MSG_TRIGGER };
-    msg_send(&m, _suit_coap_pid);
+    /* context is passed either as NULL or 0x1 for /active or /inactive */
+    char c = '0';
+
+    if (context) {
+        c += riotboot_slot_other();
+    }
+    else {
+        c += riotboot_slot_current();
+    }
+
+    return coap_reply_simple(pkt, COAP_CODE_205, buf, len,
+                             COAP_FORMAT_TEXT, (uint8_t *)&c, 1);
 }
+#endif
 
 static const coap_resource_t _subtree[] = {
 #ifdef MODULE_RIOTBOOT_SLOT
     { "/suit/slot/active", COAP_METHOD_GET, _slot_handler, NULL },
     { "/suit/slot/inactive", COAP_METHOD_GET, _slot_handler, (void *)0x1 },
 #endif
-    { "/suit/trigger", COAP_METHOD_PUT | COAP_METHOD_POST, _trigger_handler,
-      NULL },
+    { "/suit/trigger", COAP_METHOD_PUT | COAP_METHOD_POST, _trigger_handler, NULL },
     { "/suit/version", COAP_METHOD_GET, _version_handler, NULL },
 };
